# public/api-docs/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/metricas/analise/anomalias/snapshot/{id}": {
      "post": {
        "operationId": "MetricasAnomalias_detectarAnomalias",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do snapshot",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nivel_confianca",
            "required": false,
            "in": "query",
            "description": "Nível de confiança para detecção",
            "schema": {
              "enum": [
                "baixo",
                "medio",
                "alto"
              ],
              "type": "string"
            }
          },
          {
            "name": "janela_temporal",
            "required": false,
            "in": "query",
            "description": "Número de dias a considerar para o histórico",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado da detecção de anomalias"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado"
          },
          "404": {
            "description": "Snapshot não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detectar anomalias para um snapshot específico",
        "tags": [
          "Métricas"
        ]
      }
    },
    "/api/v1/metricas/analise/anomalias/batch": {
      "post": {
        "operationId": "MetricasAnomalias_detectarAnomaliasBatch",
        "parameters": [
          {
            "name": "nivel_confianca",
            "required": false,
            "in": "query",
            "description": "Nível de confiança para detecção",
            "schema": {
              "enum": [
                "baixo",
                "medio",
                "alto"
              ],
              "type": "string"
            }
          },
          {
            "name": "janela_temporal",
            "required": false,
            "in": "query",
            "description": "Número de dias a considerar para o histórico",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de anomalias detectadas"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detectar anomalias em lote para todas as métricas",
        "tags": [
          "Métricas"
        ]
      }
    },
    "/api/v1/metricas/analise/tendencias/metrica/{id}": {
      "post": {
        "operationId": "MetricasAnomalias_analisarTendencias",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da métrica",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado da análise de tendências"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado"
          },
          "404": {
            "description": "Métrica não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Analisar tendências para uma métrica",
        "tags": [
          "Métricas"
        ]
      }
    },
...
    "/api/v1/metricas/analise/{codigo}/tendencia": {
      "get": {
        "operationId": "MetricasAnalise_analisarTendencia",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataInicio",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "dataFim",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "granularidade",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Análise de tendência realizada com sucesso"
          },
          "404": {
            "description": "Métrica não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Analisa tendência de uma métrica",
        "tags": [
          "Métricas"
        ]
      }
    },
    "/api/v1/metricas/analise/{codigo}/anomalias": {
      "get": {
        "operationId": "MetricasAnalise_detectarAnomalias",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataInicio",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "dataFim",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "sensibilidade",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detecção de anomalias realizada com sucesso"
          },
          "404": {
            "description": "Métrica não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detecta anomalias em série histórica de métrica",
        "tags": [
          "Métricas"
        ]
      }
    },
    "/api/v1/metricas/analise/{codigo}/previsao": {
      "get": {
        "operationId": "MetricasAnalise_gerarPrevisao",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "horizonte",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "intervaloConfianca",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "modelo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Previsão gerada com sucesso"
          },
          "404": {
            "description": "Métrica não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gera previsão baseada em dados históricos",
        "tags": [
          "Métricas"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Gestão de Benefícios Eventuais",
...
    "version": "1.0",
    "contact": {
      "name": "Equipe SEMTAS",
      "url": "https://www.natal.rn.gov.br/semtas",
      "email": "suporte@semtas.natal.rn.gov.br"
    },
    "license": {
      "name": "Uso Interno",
      "url": "https://www.natal.rn.gov.br/semtas/licenca"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Insira o token JWT",
        "in": "header"
      }
    },
...
  }
}